projects/litreelab-studio-v2/litlab's-player-deluxe/requirements-new.txt
# Overlord PC Dashboard - Python Dependencies
# Install with: pip install -r requirements.txt

# ═══════════════════════════════════════════════════════════════
# FastAPI and Server
# ═══════════════════════════════════════════════════════════════
fastapi==0.115.0
uvicorn[standard]==0.32.0
websockets==13.1
python-multipart==0.0.17

# ═══════════════════════════════════════════════════════════════
# Database
# ═══════════════════════════════════════════════════════════════
sqlalchemy[asyncio]==2.0.36
alembic==1.14.0
aiosqlite==0.20.0
# NOTE: For PostgreSQL, install separately:
#   Windows: Install Microsoft C++ Build Tools first, then: pip install asyncpg psycopg2-binary
#   Linux/Mac: pip install asyncpg psycopg2-binary

# ═══════════════════════════════════════════════════════════════
# Configuration
# ═══════════════════════════════════════════════════════════════
pydantic[email]==2.12.0
pydantic-settings==2.13.1
python-dotenv==1.0.1
pyyaml==6.0.2

# ═══════════════════════════════════════════════════════════════
# Security
# ═══════════════════════════════════════════════════════════════
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# ═══════════════════════════════════════════════════════════════
# System Monitoring
# ═══════════════════════════════════════════════════════════════
psutil==6.1.0
nvidia-ml-py==12.535.133
structlog==25.5.0

# ═══════════════════════════════════════════════════════════════
# HTTP Client & Integrations
# ═══════════════════════════════════════════════════════════════
httpx==0.27.2
requests>=2.31.0
pure-python-adb
paramiko>=2.12.0
google-generativeai>=0.3.2

# ═══════════════════════════════════════════════════════════════
# Metrics & Observability
# ═══════════════════════════════════════════════════════════════
prometheus-client==0.21.0

# ═══════════════════════════════════════════════════════════════
# Testing
# ═══════════════════════════════════════════════════════════════
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==6.0.0

# ═══════════════════════════════════════════════════════════════
# Development Tools
# ═══════════════════════════════════════════════════════════════
black==24.10.0
isort==5.13.2
mypy==1.13.0
ruff==0.7.3

# ═══════════════════════════════════════════════════════════════
# Typing Stubs
# ═══════════════════════════════════════════════════════════════
types-requests
types-PyYAML
types-psutil

# ═══════════════════════════════════════════════════════════════
# Optional: Firebase (uncomment to enable)
# ═══════════════════════════════════════════════════════════════
# firebase-admin>=6.5.0
